refactor some stuff around
